Polishing merge tests
