+++ tests/core/pyspec/eth2spec/test/exceptions.py
@@ -1,2 +1,6 @@
 class SkippedTest(Exception):
     ...
+
+
+class BlockNotFoundException(Exception):
+    ...
